Desktop new 0.1.8.3
Az eszköz- és foglaláskezelés hiba-kezelésének és felhasználói felületének fejlesztése

Átdolgoztuk az eszközök (Eszköz) és az üzemeltetők (Üzemeltető) adatainak betöltését, jobb hiba kezeléssel, robusztus deszerializációval és egyértelműbb hibakeresési naplózással. Frissítettük az eszközök listázásának és a foglalások törlésének felhasználói felületét, beleértve a felhasználói visszajelzéseket és a megerősítő párbeszédablakokat. Az egységesség érdekében az azonosító típusokat long-ról int-re változtattuk. Az egyértelműség kedvéért átneveztük az Üzemeltető deszerializációs módszerét. Ezek a változtatások javítják a kód olvashatóságát, megbízhatóságát és a felhasználói élményt.

+++ Desktop/CyberNest_Admin/CyberNest_Admin/Eszkoz.cs
@@ -12,7 +12,7 @@ namespace CyberNest_Admin
     public partial class Eszkoz
     {
         [JsonPropertyName("id")]
-        public long Id { get; set; }
+        public int Id { get; set; }
 
         [JsonPropertyName("leiras")]
         public string Leiras { get; set; }
@@ -27,7 +27,7 @@ namespace CyberNest_Admin
         public string Hdd { get; set; }
 
         [JsonPropertyName("uzemelteto_id")]
-        public long UzemeltetoId { get; set; }
+        public int UzemeltetoId { get; set; }
 
         [JsonPropertyName("uzemelteto_nev")]
         public string UzemeltetoNev { get; set; }
